// src/interface/user.interface.ts
export interface User {
    id: string;
    name: string;
    surname: string;
    account_id: string;
    title: string;
    personal_id: string;
    department: string;
    position: string;
    priority: string;
    tel: string;
    start_date: string;
    birthday: string;
    address: string;
    medical: string;
    url: string;
    level: number;
  }